setting to ssd.pytorch

+++ ssd/layers/modules/multibox_loss.py
@@ -1,7 +1,5 @@
 # -*- coding: utf-8 -*-
-import time
 import torch
-import numpy as np
 import torch.nn as nn
 import torch.nn.functional as F
 
@@ -56,6 +54,7 @@ class MultiBoxLoss(nn.Module):
 				conf shape: torch.size(batch_size,num_priors,num_classes)
 				loc shape: torch.size(batch_size,num_priors,4)
 				priors shape: torch.size(num_priors,4)
+
 			targets (tensor): Ground truth boxes and labels for a batch,
 				shape: [batch_size,num_objs,5] (last idx is the label).
 		"""
@@ -102,7 +101,6 @@ class MultiBoxLoss(nn.Module):
 
 		# Hard Negative Mining
 		loss_c[pos.view(-1, 1)] = 0  # filter out pos boxes for now
-		#loss_c[pos] = 0  # filter out pos boxes for now
 		loss_c = loss_c.view(num, -1)
 		_, loss_idx = loss_c.sort(1, descending=True)
 		_, idx_rank = loss_idx.sort(1)
@@ -119,8 +117,7 @@ class MultiBoxLoss(nn.Module):
 
 		# Sum of losses: L(x,c,l,g) = (Lconf(x, c) + αLloc(x,l,g)) / N
 
-		N = num_pos.data.sum()
-		N = N.float()
+		N = num_pos.data.sum().float()
 		loss_l /= N
 		loss_c /= N
 		return loss_l, loss_c
